[Add] 매니저 및 데이터 스크립트 생성
