+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Data/DataHandler.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db08e6d9a21383f479ebda8783acef60
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
