+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Data/DataHandler/BoardPlayerData.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class BoardPlayerData 
+{
+    
+}
