+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Data/DataHandler/BoardPlayerData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fea7069d0de5f4141886ebad13f57d25
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
