+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Data/DataHandler/MiniGameData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9f2782563c9fc9e4f9b5d934c4295dd0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
