+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Data/DataHandler/PlayerInstanceData.cs
@@ -1,0 +1,7 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerInstanceData
+{
+}
