+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Managers/BoardGameManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 62328493894a51b45ad43f6ee0f71500
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
