+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Managers/DataManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1d43b80f56ca7fb42bb92acde57eaff7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
