+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Managers/FrameWorkManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f55773ab533477448cd48ca3cb7044d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
