+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Managers/LoadManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ba334f44cb68fc54b8b385e9f7ec89f4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
