+++ Team_Immortal Sprouts_Pummel Party/Assets/Scripts/Managers/Managers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d21042f7709cde24195dbb33951eaeae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
